feat(as-3842): add validator for any-of-following controller
