+++ packages/forms-web-app/src/routes/before-you-start/index.js
@@ -1,8 +1,11 @@
 const express = require('express');
 const anyOfFollowingRouter = require('./any-of-following');
+const {
+  rules: anyOfFollowingRules,
+} = require('../../validators/before-you-start/any-of-following');
 
 const router = express.Router();
 
-router.use('/any-of-following', anyOfFollowingRouter);
+router.use('/any-of-following', anyOfFollowingRules(), anyOfFollowingRouter);
 
 module.exports = router;
