+++ packages/forms-web-app/src/validators/before-you-start/any-of-following.js
@@ -1,15 +1,9 @@
-const { body } = require('express-validator');
+const { body, checkSchema } = require('express-validator');
+const anyOfFollowingSchema = require('./any-of-following-schema');
 
-const anyOfFollowingOptions = [
-  'none_of_these',
-  'major_retail_and_services',
-  'major_travelling_and_caravan_pitches',
-  'major_general_industry_storage_warehousing',
-  'major_dwellings',
-  'a_listed_building',
-];
-
-const rules = () => body('options').notEmpty().isIn(anyOfFollowingOptions);
+const rules = () => {
+  return [checkSchema(anyOfFollowingSchema), body('option').notEmpty()];
+};
 
 module.exports = {
   rules,
